requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.100.0
fastapi-mail==1.4.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hypercorn==0.14.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.0
passlib==1.7.4
priority==2.0.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
wsproto==1.2.0
